+++ scripts/database/schema.oracle.sql
@@ -3163,8 +3163,8 @@ CREATE TABLE search_profile (
     login VARCHAR2 (200) NOT NULL,
     profile_name VARCHAR2 (200) NOT NULL,
     profile_type VARCHAR2 (30) NOT NULL,
-    profile_key VARCHAR2 (200) NOT NULL,
-    profile_value VARCHAR2 (200) NULL
+    profile_key VARCHAR2 (2000) NOT NULL,
+    profile_value VARCHAR2 (2000) NULL
 );
 BEGIN
     EXECUTE IMMEDIATE 'CREATE INDEX search_profile_login ON search_profile (login)';